requirements.txt
streamlit==1.18.1
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.0
openai==1.7.2
requests==2.31.0
python-decouple==3.8
huggingface-hub==0.20.2
st-pages==0.4.5